pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "relational-fraud-intelligence"
version = "1.0.0"
description = "Production-grade fraud investigation platform with case management, alert automation, graph analysis, and explainable risk reasoning."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.20.0",
    "alembic==1.16.5",
    "fastapi==0.135.1",
    "httpx==0.28.1",
    "huggingface_hub==1.6.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-exporter-prometheus>=0.46b0",
    "prometheus-client>=0.20.0",
    "PyJWT==2.11.0",
    "pydantic==2.12.5",
    "pydantic-settings==2.13.1",
    "psycopg[binary]==3.2.10",
    "python-multipart==0.0.20",
    "redis==5.2.1",
    "relationalai==1.0.3",
    "sqlalchemy==2.0.43",
    "uvicorn[standard]==0.41.0",
    "networkx==3.4.2",
    "scipy>=1.14.0",
]

[project.optional-dependencies]
dev = [
    "mypy==1.19.1",
    "pre-commit==4.3.0",
    "pytest==9.0.2",
    "pytest-asyncio>=0.25.0",
    "pytest-cov==7.0.0",
    "ruff==0.15.5",
]
otlp = [
    "opentelemetry-exporter-otlp-proto-grpc>=1.25.0",
]

[project.scripts]
rfi-api = "relational_fraud_intelligence.__main__:main"
rfi-manage = "relational_fraud_intelligence.manage:main"

[tool.hatch.build.targets.wheel]
packages = ["src/relational_fraud_intelligence"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = "--strict-markers --cov=relational_fraud_intelligence --cov-report=term-missing -m 'not postgres'"
markers = [
    "postgres: tests that require a real Postgres database (deselected by default)",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "UP",
    "B",
]

[tool.coverage.run]
source = ["src/relational_fraud_intelligence"]

[tool.coverage.report]
fail_under = 85
show_missing = true

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["networkx"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "opentelemetry.*",
    "prometheus_client",
]
ignore_missing_imports = true

